Decompose JSONL scanner tests
